[FE] REFACTOR: util이 아닌 부분 MemberList로 옮기기 #161
